+++ hgmd/__init__.py
@@ -1,0 +1,4 @@
+"""This app provides HGMD variant annotation.
+
+At the moment, only the ``HGMD_PUBLIC`` trac from ENSEMBL is supported.
+"""